requirements.txt


# Core Framework
streamlit>=1.46.0

# AI and Language Models
openai>=1.0.0
langchain>=0.3.26
langchain-openai>=0.3.25
langchain-community>=0.3.0
langgraph>=0.2.0

# Data Processing and Analytics
pandas>=2.3.0
plotly>=6.1.2
pydantic>=2.0.0

# Web Search
duckduckgo-search>=5.0.0

# Scheduling and Background Tasks
schedule>=1.2.0
apscheduler>=3.10.0

# Environment Management
python-dotenv>=1.0.0

# Development and Testing
pytest>=7.0.0